Reapplied index.html and Samples.txt
Seeing as github does not allow me to pull the cam images that are hosted on http protocol websites, unfortunately I am going to be unable to host the camviewer.

Feel free to open an issue if you know a workaround for this. Yes I have tried appending https to the url request and no it doesn't work.

As a workaround I have provided the code and some samples to the CameraViewer web interface, along with some sample IP addresses which are definitely not the ones that have previously outputted to success.txt

Happy timezone :D